+++ shared/src/main/scala/amf/client/remod/AMFIdGenerator.scala
@@ -1,4 +1,4 @@
-package amf.client.`new`
+package amf.client.remod
 
 import amf.core.model.domain.DomainElement
 import amf.core.utils.IdCounter
