+++ shared/src/main/scala/amf/client/remod/amfcore/plugins/PluginPriority.scala
@@ -1,0 +1,9 @@
+package amf.client.remod.amfcore.plugins
+
+sealed case class PluginPriority(priority: Int) {}
+
+object HighPriority extends PluginPriority(1)
+
+object NormalPriority extends PluginPriority(2)
+
+object LowPriority extends PluginPriority(3)
