+++ shared/src/main/scala/amf/client/remod/amfcore/plugins/parse/AMFParsePluginAdapter.scala
@@ -1,20 +1,18 @@
-package amf.client.`new`.amfcore.plugins
+package amf.client.remod.amfcore.plugins.parse
 
-import amf.client.`new`.amfcore.{AMFParsePlugin, ParsingInfo, PluginPriority}
 import amf.client.plugins.AMFDocumentPlugin
-import amf.core.{CompilerContext, Root}
+import amf.client.remod.amfcore.plugins.PluginPriority
+import amf.core.Root
 import amf.core.client.ParsingOptions
 import amf.core.errorhandling.ErrorHandler
 import amf.core.model.document.BaseUnit
-import amf.core.parser.{LibraryReference, ParsedDocument, ParserContext, RefContainer, ReferenceHandler, ReferenceKind}
-import amf.core.remote.{Platform, Vendor}
-import org.yaml.model.YNode
+import amf.core.parser.{ParserContext, ReferenceHandler}
 
 case class AMFParsePluginAdapter(plugin: AMFDocumentPlugin) extends AMFParsePlugin {
   override def parse(document: Root, ctx: ParserContext, options: ParsingOptions): Option[BaseUnit] =
     plugin.parse(document, ctx, options)
 
-  override val supportedVendors: Seq[String] = plugin.vendors
+  override val supportedVendors: Seq[String]        = plugin.vendors
   override val validVendorsToReference: Seq[String] = plugin.validVendorsToReference.map(_.name)
 
   override def referenceHandler(eh: ErrorHandler): ReferenceHandler = plugin.referenceHandler(eh)
