+++ shared/src/main/scala/amf/core/registries/AMFPluginsRegistry.scala
@@ -1,8 +1,8 @@
 package amf.core.registries
 
-import amf.client.`new`.{AMFEnvironment, BaseEnvironment}
-import amf.client.`new`.amfcore.AMFParsePlugin
+import amf.client.remod.{AMFEnvironment, BaseEnvironment}
 import amf.client.plugins._
+import amf.client.remod.amfcore.plugins.parse.AMFParsePlugin
 import amf.core.validation.AMFPayloadValidationPlugin
 
 import scala.collection.mutable
